requirements.txt
discord.py>=2.0.0
requests>=2.31.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pytz>=2023.3
ollama>=0.1.0
PyGithub>=2.1.0
Flask>=3.0.0
